requirements.txt
Keras==2.2.4
Pillow==5.3.0
matplotlib==3.0.2
Flask==1.0.2